requirements.txt
# Chamados TI - SOTRIGO
# Dependências do projeto

# Oracle Database
oracledb>=2.0.0

# GUI
customtkinter>=5.2.0
tkcalendar>=1.6.1

# Web (Flask)
flask>=3.0.0
flask-limiter>=3.5.0

# Configuração
python-dotenv>=1.0.0

# Notificações (opcional)
plyer>=2.1.0

# Export (opcional)
openpyxl>=3.1.0

# Utilitários
Pillow>=10.0.0
